Move Duration(Type) into 'internal/provider/fwprovider'.
